L2ParamFamily(ROptEst)	R Documentation

Generating function for L2ParamFamily-class

Description

Generates an object of class "L2ParamFamily".

Usage


L2ParamFamily(name, distribution = Norm(), distrSymm,
        main = 0, nuisance, trafo, param, props = character(0), 
        L2deriv = EuclRandVarList(RealRandVariable(list(function(x) {x}), 
                                                   Domain = Reals())), 
        L2derivSymm, L2derivDistr, L2derivDistrSymm, FisherInfo)

Arguments

name 	 character string: name of the family

distribution 	 object of class "Distribution": member of the family

distrSymm 	 object of class "DistributionSymmetry": symmetry of distribution.

main 	 numeric vector: main parameter

nuisance 	 numeric vector: nuisance parameter

trafo 	 matrix: transformation of the parameter

param 	 object of class "ParamFamParameter": parameter of the family

props 	 character vector: properties of the family

L2deriv 	 object of class "EuclRandVariable": L2 derivative of the family

L2derivSymm 	 object of class "FunSymmList": symmetry of the maps contained in L2deriv

L2derivDistr 	 object of class "UnivarDistrList": distribution of L2deriv

L2derivDistrSymm 	 object of class "DistrSymmList": symmetry of the distributions contained in L2derivDistr

FisherInfo 	 object of class "PosDefSymmMatrix": Fisher information of the family

Details

If name is missing, the default “L2 differentiable parametric family of probability measures” is used. In case distrSymm is missing it is set to NoSymmetry(). If param is missing, the parameter is created via main, nuisance and trafo as described in ParamFamParameter. In case L2derivSymm is missing, it is filled with an object of class FunSymmList with entries NonSymmetric(). In case L2derivDistr is missing, it is computed via imageDistr. If L2derivDistrSymm is missing, it is set to an object of class DistrSymmList with entries NoSymmetry(). In case FisherInfo is missing, it is computed from L2deriv using E.

Value

Object of class "L2ParamFamily"

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

L2ParamFamily-class

Examples


F1 <- L2ParamFamily()
plot(F1)

[Package ROptEst version 0.5.0 Index]